langchain-python/requirements.txt
langchain>=0.0
faiss-cpu
python-dotenv
fastapi
uvicorn
langchain_google_genai
langchain_core
pytest
requests
httpx
fastapi[test]